separate models from utility functionality
